clarify usage of coverage generators
It took me about two days to realize that --cov=codeclimate_test_reporter isn't supposed to be there in my case, that it's not a magical plugin that gets and sends that report, that it's just the part you want to cover. I'd generally avoid using codeclimate_test_reporter as an example, as it can be confused. If you want a specific example, why not use an existing library that's not self? :)